Add purchase list service and tests for the service applications
The purchase list service currently can only save purchase lists, but
not items inside the purchase list. Which is now a fairly easy task to
do.

The tests are integration tests. Because the entire stack is provided by
spring including an frontend-controller, object to relational mapper,
and because of time limitations, micro testing was skipped.