Additional improvements to the build chain.
This commit improves the gradle build chain significantly. It has speed-
up features, like plugin loading, target compatibility is now set to
correct value. It also allows to create an REST API Documentation from
the integration testing. This was deemed necessary, because this code
will be used as a demonstration in class.

A fact to remember, gradle sub-projects do not need a name, because the
name was given by the super-project.