Portal composition
Compose other services by portal composition. Other service renders the
page and the main service (recipe) proxies the pages to the browser.
When the browser sends requests, the main service (recipe) proxies them
to the secondary services. On the main service is no application logic,
just proxy download and upload to the secondary services.